refactor(dify-agent): replace cost model with sizing inputs
